images are segregated into ./symbols ./junk folders
The generated images for "connect" and "no_connect" are segregated into subfolders titled "junk" or "symbols".  This is necessary because the unique numeric identifier is only unique within the dataset (either symbols or junk).

For example, there is an 101_alfonso_6 in both the trainingSymbols and trainingJunk datasets... but they contain DIFFERENT data.  Therefore, "101_alfonso_6" is not a globally unique identifier.