Alternative feature vector, compact line orientations
This is using an alternative feature vector where by a sequence of angles of the lines that comprise the symbols is used instead of the image of the symbol.  This is more compact and may be more efficient to classify.  This approach is a possible alternative to using image data.