Final updates to use caching to improve performance and reduce the number of ABSENT symbols
Added lazy caching of the combinations used when generating candidate symbols from strokes.
Revised the heigh/width min-max normalization to preserve aspect ratio; this allows for the minDistance parameter to behave more accurately now that the normalized strokes for a candidate symbol are not "warped" along any axis.
During final validation of the segmentation sets, added some logic to force absent strokes to be represented in the segmentation.  This prevents excess absent strokes from appearing in the output files.